+++ apps/cli/src/commands/export/export-page.ts
@@ -1,0 +1,33 @@
+import {
+  ComponentModel,
+  componentSelectionSet,
+  Page as IPageModel,
+} from '@codelab/backend'
+import { flatMap, flatten } from 'lodash'
+import { getElementAndDescendants } from './export-element'
+
+export const getPageData = async (page: IPageModel) => {
+  const Component = await ComponentModel()
+  const elements = await getElementAndDescendants(page.rootElement.id)
+
+  const componentIds = flatMap(elements, (e) => [
+    e.component?.id,
+    e.instanceOfComponent?.id,
+  ]).filter(Boolean) as Array<string>
+
+  const components = await Component.find({
+    where: { id_IN: componentIds },
+    selectionSet: componentSelectionSet,
+  })
+
+  const componentRootIds = components.map((c) => c.rootElement.id)
+
+  const componentElements = await Promise.all(
+    componentRootIds.map((id) => getElementAndDescendants(id)),
+  )
+
+  return {
+    elements: [...elements, ...flatten(componentElements)],
+    components,
+  }
+}
