+++ apps/cli/src/commands/import/import-app.command.ts
@@ -4,9 +4,7 @@ import {
   Atom as IAtomModel,
   AtomModel,
   Component as IComponentModel,
-  ComponentModel,
   Element as IElementModel,
-  ElementModel,
   Page as IPageModel,
   UserModel,
 } from '@codelab/backend'
@@ -14,7 +12,10 @@ import { config } from 'dotenv'
 import fs from 'fs'
 import * as inquirer from 'inquirer'
 import { flatMap } from 'lodash'
+import path from 'path'
 import yargs, { CommandModule } from 'yargs'
+import { importComponent } from './import-component'
+import { importElementInitial, updateImportedElement } from './import-element'
 
 interface PagePack {
   page: IPageModel
@@ -27,7 +28,7 @@ export const importAppCommand: CommandModule<any, any> = {
   handler: async ({ filePath }) => {
     config({ path: `${process.cwd()}/.env` })
 
-    const json = fs.readFileSync(filePath, 'utf8')
+    const json = fs.readFileSync(path.resolve('data', filePath), 'utf8')
     const Apps = await AppModel()
     const Users = await UserModel()
     const allUsers = await Users.find()
@@ -139,129 +140,4 @@ const validate = async (pages: Array<PagePack>) => {
   if (notFound.length) {
     throw new Error(`Can't find Atoms with ids "${notFound.join(', ')}"`)
   }
-}
-
-// Creates the element without prop map bindings and without parent/children connections
-const importElementInitial = async (
-  element: IElementModel,
-  idMap: Map<string, string>,
-): Promise<IElementModel> => {
-  const Elements = await ElementModel()
-
-  const {
-    elements: [newElement],
-  } = await Elements.create({
-    input: [
-      {
-        name: element.name,
-        css: element.css,
-        atom: element.atom
-          ? { connect: { where: { node: { id: element.atom.id } } } }
-          : undefined,
-        component: element.component
-          ? {
-              connect: {
-                where: { node: { id: idMap.get(element.component.id) } },
-              },
-            }
-          : undefined,
-        instanceOfComponent: element.instanceOfComponent
-          ? {
-              connect: {
-                where: { node: { id: element.instanceOfComponent.id } },
-              },
-            }
-          : undefined,
-        props: element.props
-          ? {
-              create: { node: { data: element.props.data } },
-            }
-          : undefined,
-        propTransformationJs: element.propTransformationJs,
-        renderForEachPropKey: element.renderForEachPropKey,
-        renderIfPropKey: element.renderIfPropKey,
-      },
-    ],
-  })
-
-  return newElement
-}
-
-// Updates the imported element with prop map bindings, parent/children connections and props after we have
-// imported all the elements, so we can reference them
-const updateImportedElement = async (
-  element: IElementModel,
-  idMap: Map<string, string>,
-): Promise<void> => {
-  const Elements = await ElementModel()
-
-  if (element.props) {
-    // replace all references in props
-    for (const [key, value] of idMap.entries()) {
-      element.props.data = element.props.data.replace(
-        new RegExp(key, 'g'),
-        value,
-      )
-    }
-  }
-
-  await Elements.update({
-    where: { id: idMap.get(element.id) },
-    update: {
-      parentElement: element.parentElement
-        ? {
-            disconnect: { where: {} },
-            connect: {
-              edge: { order: element.parentElementConnection?.edges[0].order },
-              where: { node: { id: idMap.get(element.parentElement.id) } },
-            },
-          }
-        : undefined,
-      props: element.props
-        ? {
-            update: { node: { data: element.props.data } },
-          }
-        : undefined,
-      propMapBindings: element.propMapBindings.map((pmb) => ({
-        create: [
-          {
-            node: {
-              sourceKey: pmb.sourceKey,
-              targetKey: pmb.targetKey,
-              targetElement: pmb.targetElement
-                ? {
-                    connect: {
-                      where: {
-                        node: { id: idMap.get(pmb.targetElement.id) },
-                      },
-                    },
-                  }
-                : undefined,
-            },
-          },
-        ],
-      })),
-    },
-  })
-}
-
-const importComponent = async (
-  component: IComponentModel,
-  selectedUser: string,
-): Promise<IComponentModel> => {
-  const Components = await ComponentModel()
-
-  const {
-    components: [newComponent],
-  } = await Components.create({
-    input: [
-      {
-        name: component.name,
-        owner: { connect: { where: { node: { id: selectedUser } } } },
-        rootElement: { create: { node: { name: '' } } },
-      },
-    ],
-  })
-
-  return newComponent
 }
