+++ libs/backend/src/selectionSets/appSelectionSet.ts
@@ -1,0 +1,1 @@
+export const appSelectionSet = `{ id, name, rootProviderElement { id } }`
