+++ libs/backend/src/selectionSets/index.ts
@@ -1,3 +1,4 @@
+export * from './appSelectionSet'
 export * from './componentSelectionSet'
 export * from './elementSelectionSet'
 export * from './pageSelectionSet'
